fix: Next.js auth route
